Add timers for llvm codegen
Summary: This will give us a rough idea of how much time we spend in the
various parts of the llvm backend. I considered adding a mode to Timer that
aborts if we punt by throwing an exception, but it looks like it's not possible
to get that 100% correct until std::uncaught_exceptions() in C++17.

Reviewed By: @jasone

Differential Revision: D1739445

Signature: t1:1739445:1418674406:e61081656e29da64c2ddabd98477cf0eab099bb7