Enable dead code warnings in typing/
Summary: @pikatchu said it was worth keeping the SubstPos code around,
because we ideally should re-implement it at some point. It was disabled
because it was buggy, but working right it would still be a useful
feature. To avoid OCaml yelling at us for unused code, we declare the
relevant functions in `mli`s.

Reviewed By: @elgenie

Differential Revision: D1742207

Signature: t1:1742207:1418746019:22aebbc9f8e6e099df3b6a3258206846a6ff0cb7