Don't use orqim on a 16-bit value
Summary: More code that was little-endian specific and getting messed up by the
llvm optimizer. In this case it was combining a write to an ObjectData's
o_attribute and m_count fields, turning the count into garbage and causing the
object to leak. I did some history digging and as far as I can tell, the only
reason this wasn't an orwim to begin with was because it was copied from some
TranslatorX64 code I wrote back in the dark days of the assembler, when we
didn't have a 16-bit version of or.

Reviewed By: @jasone

Differential Revision: D1742340