Faster UnitCache
Summary: This tries to put things back the way they used to be. The
assumption is that in the steady state, all the files are already
in the UnitCache, so look them up there first. If we don't find
them there, then we do the slow things like realpath, and creating
StaticStrings. If the realpath is different from the request path,
we store the unit under both paths.

Reviewed By: @jdelong

Differential Revision: D1744202