Remove code for eliminating DefInlineFP w/o eliminating the callee's stack
Summary: This was needed for eliminating the outer actrec when inlining
functions that call other functions, but this didn't work (knownPrologue
stuff).  There shouldn't be a performance implication to just remove it right
now, since it shouldn't be happening (we should only be eliminating the
DefInlineFP in situations that the first ReDefSP would've been dce'd also), but
I'm putting it in its own diff to double check.

Reviewed By: @mhahnenberg

Differential Revision: D1747541

Signature: t1:1747541:1419377102:164cecd81bb7167e539f0fb67d118c18f21b0fd6