Fix Memcached::setOption for OPT_LIBKETAMA_COMPATIBLE
Summary: The current implementation is missing an if-statement which causes
the option to be unset even when a truthy value is passed in. The comment
directly above this block describes the correct behavior. See the original PHP
implementation for reference: http://goo.gl/7p7WqP. A new constant
Memcached::DISTRIBUTION_CONSISTENT_KETAMA is added for the sake of testing.

Closes https://github.com/facebook/hhvm/pull/4364

Reviewed By: @paulbiss

Differential Revision: D1710109

Signature: t1:1710109:1418880558:9a97113ad941927c06ca0ed8389f2208b18eca86