Actually get the realpath first thing when searching for a file.
Summary: D1729338 fixed a subtle bug, but not completely. We didn't handle the case where
the test runner was run in the exact same directory as the test in question. If we
did, it would still come in as `.` if `dirname()` was used beforehand, which is
probably not desired behavior. So move the call to `realpath()` to before the
`while()` loop and all cases should now be handled.

{sync, type="child", parent="internal", parentrevid="1742941", parentrevfbid="866765913387291", parentdiffid="5964829"}

Reviewed By: @paulbiss

Differential Revision: D1742941

Signature: t1:1742941:1418878079:1499d0d009589f7a47a1f1a04e08619c2813b673