reorder shm_put_var operations
Summary: Reordering the shm_put_var operation to serialize the object before
writing to memory aviods a segfault and possible arbitrary memory write.
Previously the shared memory location was found before the object was
serialized and it is possible to detach the shm location during serialization.
This would have shm_put_data write to memory using a shm handle that doesn't
exist.

CVE-2010-1861

Reviewed By: @markw65

Differential Revision: D1694899