Relax boxed prediction invariants enough to remove ConstrainBoxedFunc
Summary: The ConstrainBoxedFunc callbacks used to be used to deal with
inner-ref types, but now they effectively just turn BoxedCells into
BoxedInitCells, which was needed because frame-state required slots with boxed
predictions to be BoxedInitCell.  This diff tracked down a few places that
should be BoxedInitCell to try to keep that, but because of the way we deal
with static locals in closures the invariant is actually too hard to keep, so
this changes it to only require that it maybe(Type::BoxedCell).

Reviewed By: @alexmalyshev

Differential Revision: D1753392