Fix lessphp repo location
Summary: cgcai's clone no longer exists. Found the PR which just added a
phpunit.xml and did it from an override instead. Longer term we should update
the versino that we test, but this is a quick fix so that we can actually
run the current test suite

{sync, type="child", parent="internal", parentrevid="1759082", parentrevfbid="322383214631507", parentdiffid="6042491"}

Reviewed By: @JoelMarcey

Differential Revision: D1759082

Signature: t1:1759082:1419877658:03c350d855de0ab3474536977aa1127c142244e8