Fix a bug and clarify something important in store-elim
Summary: While adding stack slots to this, I started getting confused why it
was removing them before I updated ReturnEffects to kill the stack slots (but
after I added that to EndCatch).  It's because ReturnEffects was only killing
the frame locals, not adding an upward exposed uses of everything else, which
is essentially a bug.  The bug doesn't cause a problem for IR we currently
generate, because we always have an EndCatch associated with the return (for
the surprise check block) which will add everything to the gen set, and only
the locals get blocked by memory_effect's knowledge of ReturnHook's
ar->setLocalsDecRefd stuff, which are correctly killed by ReturnEffects also.
But it's still wrong.

Also add a comment clarifying why liveOut sets are not initialized to true in
the exit blocks, which is probably unusual and seemed like the bug at first but
isn't.

Reviewed By: @elgenie

Differential Revision: D1760123

Signature: t1:1760123:1419903299:a870ae6209108e858484fb0851b1ce2b6f6bd379