Add vasm pass for removing trivial nops before we do anything else
Summary: This will make D1752689 not break optimizeExits, and maybe it'd
worthwhile because it just generally makes it legal to generate things like
self-leas with no penalty.  Although if we're going to do something like this
first, maybe it should be the full-blown dead code pass?  (If we decide it's
not worthwhile we can just remove it later---for now it'll probably make it
easier for me to measure the perf impact of that diff by itself.)

Reviewed By: @edwinsmith

Differential Revision: D1762062

Signature: t1:1762062:1420123008:e0cda8664e2bf854533ce4393f7df6497c6e5d8d