Implement AlwaysEscapeLog runtime option
Summary: Previously, AlwaysEscapeLog was never checked effectively making the default "true", despite the fact that it was listed as "false" in runtime-option.cpp. After checking AlwaysEscapeLog I changed the default behavior to true, to match how it behaved by default before this patch.

It's important to note that if AlwaysEscapeLog is set to false this implementation will cause logs to *never* be escaped. If that was not the intended behavior of this option, I can amend it to match what it should do.

fixes #4437
Closes https://github.com/facebook/hhvm/pull/4415

Reviewed By: @sgolemon

Differential Revision: D1727958

Signature: t1:1727958:1420498150:6b5688bcfe6a7b1a86309f8b3bee0934bd110c3d