Support EnableArgsInBacktraces in memory_effects
Summary: load-elim didn't really need to bail for this, but I had
memory_effects asserting about it.  OTOH, the may-load information is manditory
for store-elim or it could eliminate stores to argument locals that aren't
really dead.

Doesn't try to be precise about it: if an instruction can re-enter, just assume
it could load any frame local when this flag is on.

Reviewed By: @edwinsmith

Differential Revision: D1766340