Have server explicitly signal when build is finished
Summary: Also ensure that the server-client connection is closed if an
exception is thrown from the build hook. (Previously, we would do this
for exceptions thrown in the initial build step, but not in the build
hook.) Together with the explicit `BUILD_FINISHED` message, this enables
us to detect unexpected exceptions thrown during build.

Reviewed By: @elgenie, @jwatzman

Differential Revision: D1767811

Signature: t1:1767811:1420588911:c248175418fe8538a5f1f28f3a42289b7e058fcd