Add compile time option to disable thrift extension
Summary: The original aim was to make this a runtime option, but there doesn't seem to be a clean way to do it. Each extension constructor adds the instance to s_registered_extensions, which is later iterated and queried in-the-raw in several places.
Closes https://github.com/facebook/hhvm/pull/4585

Reviewed By: @JoelMarcey

Differential Revision: D1767122

Signature: t1:1767122:1420665326:70e34284f93a2bed50729df786865fdc7ed697ca