Fix some problems with HHIR generic memory ops that take C(Int) sources
Summary: The pointee() function was ignoring possible offsets, so just refactor
things so we always use offset zero with these.  Also add a DbgTrashMem.  (I
noticed this problem partially through adding that.)  The packedarray functions
were handling it properly.  LdElem and StElem are used with non-constant
integer offsets---I want to replace them with a new typed-lea and StMem/LdMem,
but for this diff just make them not incorrect they way they are.

Reviewed By: @alexmalyshev

Differential Revision: D1772985