Fix attribution of CPU time spent in the asio scheduler
Summary: The cost of the asio scheduler may be nontrivial, especially the cost
of unserialization of result of ExternalThreadEventWaitHandle.

If a Xenon event occurred before a function entry, we attributed the
cost to its caller. This is wrong when the function was entered by
resuming from asio scheduler, as it attributes the cost of
unserialization to the caller of an async function that awaited the I/O
operation.

Let's fix this by attributing the cost to the function awaiting the I/O.

Ideally, the cost would be attributed to the extension function, but at
the moment of unserialization, the frame of the extension function no
longer exists (possibly for very long time).

Reviewed By: @billf

Differential Revision: D1773617

Signature: t1:1773617:1420765732:1afa7613b4e0c170024fc4fe6f69c49b72165b30