Finalize Type Const syntax
Summary: This is my proposed final syntax for the Type Constant feature. It
essentially borrows from both declaring constants in classes, as well as
declaring typedefs.

Essientially it is

[abstract] const NAME [as CONSTRAINT_TYPE] [= ASSIGNED_TYPE]

If a type constant has a CONSTRAINT_TYPE then it is allowed to be
overriden by it's child with a type that satisifies the constraint.
The ASSIGNED_TYPE can be thought as the runtime type.

This is similar to opaque types where you write

newtype ID as int = int

Reviewed By: @jwatzman

Differential Revision: D1773668