Fix bizarre memory corruption on exit
Summary: - Some things, eg the UnitCache can send items to the treadmill
   when they're destroyed.
 - The treadmill may have been destroyed first.
 - Adding an entry to a destroyed std::list turns out to be safe if the
   list was empty when it was destroyed
 - The treadmill is almost always empty when its destroyed, because
   all requests have finished
 - FreeStubTrigger can add itself back to the treadmill if something else
   has the write lease; in that case, treadmill is non-empty when destroyed

Even then, all that happened was that a pointer was written into a freed
block of memory that would never be touched again. But tcmalloc checks
freed memory from time to time, and asserted if it found a problem.

Reviewed By: @jasone

Differential Revision: D1775126