Add ExecutionContext::{invokeMethod,invokeMethodV}
Summary: It's meant as a more convenient way of running a method on an object, when the
arguments are known and there isn't any funny business going on with contexts or
__call. Inspection of most of the calls of invokeFuncFew show this is the common
case.

invokeMethodV is the same as invokeMethod, but returns a Variant instead of a
TypedValue. There were enough callsites where the user made a Variant, then
passed it to invokeFuncFew via a TypedValue* that I figured I should add it too.

Note the ObjectData* and const Func* arguments are reversed from their
respective order in invokeFuncFew. I figured the tuple "(Obj, Meth)" made more
sense than "(Meth, Obj)".

Reviewed By: @paulbiss

Differential Revision: D1746147

Signature: t1:1746147:1418850387:8ff930839862c46127056078380c860faa4e4c09