Improve position info in member-not-found error
Summary: For our "Declaration of X is here", fix the position to point to the declaration site, not the site of object instantiation.

Add a separate error message for why we believe the object to be a given type. The position for this error message uses `Reason.to_string` instead of merely the position of the class type, so that we get better error messages when the type is due to usage of `$this`, `new static()`, etc.

Reviewed By: @elgenie

Differential Revision: D1770937