Fix mysqli segfault
Summary: When we moved to a typed `MySqlResource` in D1729488, one of the good Zend tests
started failing with a segfault:

```
test/zend/good/ext/mysqli/tests/bug28817.php
```

I believe this was because we could possibly return a `nullptr` from `getTyped` and
when it was `nullptr` we would still be trying to make a call to `mysql()`. So now
we do explicit checks for this.

{sync, type="child", parent="internal", parentrevid="1771416", parentrevfbid="801601039906010", parentdiffid="6108137"}

Reviewed By: @mxw

Differential Revision: D1771416

Signature: t1:1771416:1420706894:ae817e4357df7dd5f332647951748e2a15f5cb60