Add --no-load option, make load script path relative
Summary: Adding --no-load makes it easier to check if a bug may be due
to saved state.

Also allow the load script path to be relative to the repository root,
so that we can ensure that the script's version is compatible with
hh_server.

Also, since the previous implementation did not have versioning, and the
current implementation changes the script output, I've renamed the
Python load script so that earlier versions of hh_server would not
attempt to run and parse the output of the latest script.

Reviewed By: @jwatzman

Differential Revision: D1769672

Signature: t1:1769672:1420753893:25f67d85fda67a5006c62cf1aa2d95752a60548a