Allow array-get syntax for KeyedContainer
Summary: Also remove Indexish from our mock interfaces in
hh_single_type_check, and have `idx` take a `KeyedContainer` as its
argument.

The eventual goal is to remove Indexish entirely.

This also improves the "this is not a container" error message -- people
tended to think that the indexed object needs to be a `Container`.

Reviewed By: @elgenie

Differential Revision: D1766126

Signature: t1:1766126:1421281844:262ed6943ccbe73e30d634002cea802bd4cd6576