Use separate lexer for comments
Summary: I think this is the right (and simpler) way to do things, and
it removes the need to handle things like '**' in a comment.

Funnily enough, it also revealed that the test case that I introduced to
check for '**' had another bug in it -- the second apostrophe in `'**'` was
missing in the formatted output (no doubt due to some other unhandled
token issues).

Reviewed By: @jwatzman

Differential Revision: D1777144

Signature: t1:1777144:1421199620:900929dad5567cff9b172a9ea3c3f6e2c8c32b12