format: fix duplication of tokens in diff mode
Summary: Since --diff mode selectively interleaves newline-delimited blocks of
original source with blocks of formatted source, we must ensure that the
newlines in the original and formatted source match up to avoid
duplicated and / or missing tokens.

There might conceivably be a cleaner fix that has TextBlocks ensure that
its output indeed maintains the invariant that Block delimiters align
with newlines in the original source, rather than relying on its input
to be up to spec... but I think this is good enough for now.

This might make `hh_format --diff` a little more aggressive about
reformatting unmodified lines, and we could fix that in a future diff
by inserting more Blocks at sensible points. But this diff should at
least ensure correctness.

Reviewed By: @jwatzman, @elgenie

Differential Revision: D1779218