Make ObjectData::propImpl easier to understand
Summary: This nightmare would take a TypedValue*& that was initially bound to &tvScratch,
but might be later set to either &tvRef, &init_null_variant, or the address of
the property's TypedValue. It also had if-statements nested 5 levels deep.

Its tvRef parameter is now a TypedValue*, since we're going to take the address
of it anyway, it now takes an explicit tvScratch as a TypedValue*, and it now
returns the address where the property ended up.

Delete ObjectData::invokeGetProp(), that was just adding to the confusion.

Reviewed By: @swtaarrs

Differential Revision: D1774218

Signature: t1:1774218:1421195332:78bc107fb4cdb32f8723e36aeaed4c89c3c01fdb