Fix SmartPtr move constructor on Variant.
Summary: The SmartPtr move constructor on Variant was wrong. It would detach the pointer and pass it along to the regular Variant ctor which would increment the refcount. This would leave the pointed to value with an extra refcount.

I added some new Variant ctors that don't bump of the refcount and just attach the passed in pointer.  I also added a unit test.

Reviewed By: @swtaarrs

Differential Revision: D1780830