Format: Use header detection logic in the parser
Summary: It is kind of redundant to have Format_hack figure out the mode
of the file when the parser has already done it. Instead, we expose the
file mode in the return value of `Parser_hack.program` and have the
formatter use that value.

This exposed a bug where the formatter wasn't handling a decl alias.
Aside from that, there should be no change in functionality.

I'm intending to allow the formatter to be run on PHP / decl files via a
command-line flag, and using the parser's mode logic makes that easier.
That's also why I renamed the `Php_or_decl` exception to
`Disabled_mode`.

Reviewed By: @jwatzman

Differential Revision: D1787849

Signature: t1:1787849:1421450259:122724031d1b2b4f1be64ee97288cb6e4afa93af