Add option to format PHP and decl files
Summary: Just because we can't parse them doesn't mean we can't unparse them! Of
course, it means that we cannot safely make certain assumptions about
the AST and so the output may be syntactically wrong... but as long as
the user double-checks the result it's still a useful tool.

(based on a conversation with@periodic1236)

Reviewed By: @jwatzman

Differential Revision: D1787855

Signature: t1:1787855:1421451156:8859823f5a0e910693b3cf3828aab173c77ea654