Make gd-jpeg compatible with jpeg-9
Summary: The "safeboolean" logic in gd_jpeg.c was well intentioned, but actually wrong.
Once jmorecfg.h has been included (which it has been by this point),
we're guaranteed that `boolean` it available from /somewhere/,
HAVE_BOOLEAN just indicates whether it's a provided intrinsic
or a libjpeg defined typedef.

The existing pattern worked up to jpeg-8 because `boolean`,
whether an intinsic or a typedef, was always compatible with
an int (in fact, when not an intrinsic, it was typedef'd to int),
but as of jpeg-9, it's an enum, which can't be implicitly converted.

Closes #4655

{sync, type="parent", child="external", childrevid="31695", childdiffid="165645", parentdiffid="6199778"}

Reviewed By: @JoelMarcey

Differential Revision: D1787155

Signature: t1:1787155:1421437730:c0070b398eb6dd2bee74fac9d55f29da97988ffb