Replace member helper argument macros with struct MemberState
Summary: Pass around a struct reference instead of a bunch of individual
variables. sink conversion of MemberState.ref and ref2 from
Variant& to TypedValue& to the places we excape to utility code,
instead of when calling getHelperPre() or memberHelperPre().

Simplify getHelperPost by having it return tvRet*, and eliminating
the saveResult template param; we only called it in two places,
and only one of them needed the saveResult logic. Now, getHelperPost
and setHelperPost both just take ndiscard, and no other state.
Neither of them accessed the tvRef[2] args we were passing in.

Reviewed By: @jdelong, @markw65

Differential Revision: D1802261