The Box IR opcode should take a Cell instead of Gen.
Summary: The ir spec doesn't match the implementation here. We only emit a Box in one
function (emitVGetL), and it's only emitted when the input is known to be
unboxed. Furthermore, the C++ helper we emit in codegen doesn't even check if
the input is boxed or not, it will always box it.

Reviewed By: @edwinsmith

Differential Revision: D1788633

Signature: t1:1788633:1421781691:9e2a03ed704f1cac8fc35184488734c4c4ae3002