modify hash_string* code to use CRC32
Summary: When USE_SSECRC flag is defined (always for x86_64 unless NO_SSECRC is defined),
we try to use CRC32 hash functions.
A cpuid check is done, if SSE4.2 is not available, the fallback implementation uses MururHash.

If NO_SSECRC is defined, we won't have the overhead of cpuid checking and generates
more compact code.

Reviewed By: @markw65

Differential Revision: D1776398