zend initialization settings now work better.
The const zend_ini_entry received from the PHP_INI_BEGIN..PHP_INI_END
data structure in the zend extension's source code is now copied
and associated with the system or user (per-request) CallbackMap.
This association is done with a new intermediate class UserIniData,
and mechanisms for registering a factory to produce (sub)classes of
UserIniData depending on the fundamental underlying semantics of the ini
setting.

The ini setting mechanisms in the zend compat layer have an instance
of a subclass of UserIniData to hold a pointer to the copy of the
zend_ini_entry, and that copy is freed when the request finishes or the
system finishes, perhaps with some "don't care" elision on a fast path
towards exit.

The string value assigned to a zend ini setting, as for example from
"ini_set", is now saved using estrndup, so that it can be regurgitated
directly when the zend ini setting is to be gotten with "ini_get". The
string is implicitly released at request finish time, or explictly
released on the next call to ini_set.

This change removes a code smell wherein the const zend_ini_entry was
const_cast'ed to remove the const.

There's still code smell regarding the encoding of PHP_INI_ALL.  The bit
encoding of PHP_INI_ALL in zend differs from the encoding used HHVM,
and since the bitset from zend is presented as is to hhvm, bit-wise
operations on PHP_INI_ALL are potentially incorrect.

This change also rewrites for clarity the critical code that
differentiates between user (per-request) and system settings. There's a
slight change in semantics: the new code will stickly keep the setting
in user space if it is already there. The rewrite also fixes a bug(s)
related to the encoding mismatch of PHP_INI_ALL between zend and hhvm.

There was a bug in the optimization of ini_get, wherein a call to
ini_get with a constant argument referencing a user setting would get
incorrectly treated as if it were a system setting. This bug prevented
ini_get from seeing the previous ini_set to the same setting.

This commit adds a single test case of ini_set/ini_get on the
ezc_test.integer and ezc_test.string.

Known failures of ini settings associated with HNI extensions (such as
mongo or mysql) have not been addressed in this commit, which focuses
solely on improving the zend extension layer.