Fix bug involving inlining and bytecode control flow
Summary: With bytecode control flow, a DefSP is inserted at merge points.  But
if inlining happens in one of the branches, that makes the ReDefSP at
the end of the inlined body appear unused, and it can be DCE'd away,
causing rbx not to be adjusted on that path.  This diff fixes the
problem by replacing that ReDefSP with a ResetSP, which basically does
the same thing but is not a candidate to DCE.

Reviewed By: @jdelong

Differential Revision: D1801432