Replace the DECODE macros with functions.
Summary: This gets rid of the remaining weird macros used in the interpreter.
In addition to the DECODE macros, DECODE_ITER_LIST and FREE_ITER_LIST
were only called in one place, so I inlined them.

Added an SpropState struct and replaced SPROP_OP_PRELUDE/POSTLUDE
with macros that mutate the state struct. Rewrote GETS as a template
function.

Reviewed By: @jdelong, @markw65

Differential Revision: D1802269