Fix how the HardReturnTypehints optimization deals with interfaces
Summary: As HHBBC does iterative analysis, it aims to uphold an invariant that return
types will only be refined and never be widened. When HardReturnTypehints is
enabled, there are certain code patterns that make it possible to violate
this invariant when an interface return typehint is involved.

This tweaks the HardReturnTypehints optimization to be more conservative
with interface return typehints to avoid the problem.

Reviewed By: @jdelong

Differential Revision: D1805456