Move code that emits guards out of translateRegion
Summary: This is the first step to make translateRegion more readable.  I was
surprised that the guard-emitting code was inside the loop that
translates each bytecode instruction in a block, which is weird
because all guards are emitted before the first instruction.  So I
pulled that code out into its own function, and it's now called before
that loop.

Also, getInputs() didn't need the SrcKey parameter, so I killed it.

Reviewed By: @alexmalyshev

Differential Revision: D1806194