Simplify hhi root handling
Summary: The only consumer of get_hhi_root exits the program if the hhi
root is not found, so we might as well just have `get_hhi_root` handle
that error case. This avoids having an awkward double-option-type. The
fact that `root` can take the value `Some None` is particularly
confusing.

Reviewed By: @elgenie

Differential Revision: D1803120