Make access log %h behave more like Apache
Summary: The access log format appears to be Apache like -- but doesn't behave  like Apache when hostname lookups are off -- it should log IP address in this case, this accomplishes that.

This is a dupe of https://github.com/facebook/hhvm/pull/3577 which was closed the first time due to CLA confusion.

We've been using this for 5 months and it dramatically simplifies debugging.
Closes https://github.com/facebook/hhvm/pull/4721

Reviewed By: @markw65

Differential Revision: D1807076

Pulled By: @jwatzman