Eliminate unsoundness of 'this' references to typeconst
Summary: Currently this references to type constants are not sound. When a this reference to a type constant is in a contravariant position it is not safe to unify with another this reference of the type constant in general as shown here D1739700#16

The solution I am suggesting is making this references to type constants Path Dependent Types like type members are in Scala (discussed partially in this paper http://lampwww.epfl.ch/~amin/dot/fool.pdf). To explain how this is done let's start with how a this::T is translated.

```
class C {
  const type T as arraykey = int;
}
```

Given the above class, ##this::T## will be translated to

```
(Tgeneric
  ("[unresolved]", Some (Tapply "C"))
 , "T"
 )
```

In words we create a generic type "[unresolved]" with a constraint of "C". During local typing (typing a block/function body) we need to rename "[unresolved]" based on how it is called. Here are the high level rules:

1. Static members/methods are always referenced with a class_id. Substitute "[unresolved]" with (class_id.to_string, class_id.type)

2. We do the same thing as above when calling a constructor

3. Instance members/methods, are referenced with an expression. Substitute "[unresolved]" with (expr.to_string, expr.type). For all expressions (except This) we generate temporary identifiers

We also need to do some magic for a method definition when typing its body.

1. The return type of the method we are typing needs to substitute "[unresolved]" to ("this", Tgeneric "this", Env.get_self)

2. For the input params of methods "[unresolved]" is renamed to "$this" for non-static methods, or "static" in static methods

The idea is that the "this" type should mean the Late Static Bound type.

This is a bit complex so it might be worth going over this offline. In a follow up diff I plan on limiting the use of this referenced type
constants to only covariant positions except for constructors and static methods where we also allow them in contra variant positions. This isn't to limit unsoundness, instead to avoid programmers from writing methods that the current approach would never allow to be called. The error messages that result are confusing so I'll rather disable it in case we can improve this approach in the future to be more expressive

Reviewed By: @jwatzman

Differential Revision: D1809761