Rework Type Consts to not use type defs
Summary: This diff reworks the type constant implementation to no longer create
type defs as part of its internal representation. After working with
these for a while, I felt that type defs didn't add much to the solution
besides confusion.

Instead I define a typeconst_type in the typing phase that holds the
information we need, primarily a name, a constraint, and the assigned
type. I also simplified the handling of static resolution and
typing_taccess overall to utilize instantiation. Instead of carrying
around a class_id, I translate the root of a type access to a type.

For static this type is a Tgeneric "static" with a constraint of ##this##.
With this representation we do not need to add the hack of switching out
static with ##this## in typing_instantiate.

This diff sets the stage for fixing the unsoundness issue discussed in
D1739700. This will be done in a follow up diff

Reviewed By: @jwatzman

Differential Revision: D1804177