Remove objprof_start
Summary: Now that we're using the heap parsing API (thanks Ed!) to implement
objprof, we don't have to turn it on/off, because there's no more tracking
data structure needed.

Before this goes to release we'll need a www diff to move the sampling logic
from around `objprof_start` to `objprof_get_data`.

Reviewed By: @alexmalyshev

Differential Revision: D1815763