Fix use-after-free after FastCGI protocol errors
Summary: If there was any sort of FastCGI protocol error at all,
`FastCGISession` would call `FastCGIConnection::onSessionClose` which
would deallocate the connection and session objects. This lead to a
myriad of use-after-free bugs; for example, the session might keep
reading buffered data even though it had been deallocated (partly
because it was missing a check for the phase having become invalid,
though if you are deallocated under yourself, even checking that phase
is use-after-free), and there might be transports representing incoming
data that are still in the queue to be processed, now with a dangling
pointer to the connection.

Instead, just don't call `onSessionClose`. With the proper checks now
in place for the phase being set to invalid, everything will largely
become a no-op, until the queue of transport objects drains and the
connection manager tells us `onSessionClose` of its own accord.

Also replaces the direct calls to `delete this` to `destroy()` to play
nicely with the `DestructorGuard` in `readDataAvailable`. I'm less sure
about this one, but it looks like the original author was trying to
prevent more use-after-free this way, except did it wrong.

This entire body of code is a gigantic mess, with random pointers,
shared pointers, etc flying all around, no clear lines of ownership, etc
etc. Make no mistake, this diff is just pasting over one particularly
bad problem, and there are almost certainly many more lurking around.

{sync, type="parent", child="external", childrevid="32583", childdiffid="168897", parentdiffid="6331722"}

Reviewed By: @paulbiss

Differential Revision: D1816135