Stop the failed link warnings on shutdown
Summary: When you stop a server with LightProcesses,
you tend to get a lot of messages about failing to create
a link.

It turns out that the first time a process tries to write to
the log, it will (depending on runtime options) try to backup
the existing symlink, then delete it then create a new one
pointing to the log file.

Often, the first message a LightProcess will log is the one
saying that its exiting, so on exit several processes all
race to move or unlink, and the create the same symlink, even
though it already exists. Naturally, some of them fail to create
the symlink.

This just checks to see if it exists before going through the create
path. There's still a possibility of a race, but generally, there
will only be one process trying to create it initially, so its not
a problem.

Reviewed By: @paulbiss

Differential Revision: D1816362