Avoid duplicating deprecated errors
Summary: We always pass the return value of `fun_type_of_id` to `call_`, so we
only need to do the deprecated check in `call_`.

The column numbers in the error messages changed somewhat because they
now correspond to the position of the function call, instead of just the
function id. (Previously, both would be reported by hh_client, but
hh_single_type_check would only print the first one.)

I also changed the fun_id test to array_map instead of call_user_func
because the latter activates the `call_` code path, and we want to make
sure an uncalled fun_id also raises an error. (We still duplicate errors
in the case where a fun_id is called by `call_user_func`, but I think
that is an edge case we do not need to worry about.)

Reviewed By: @jwatzman

Differential Revision: D1811946