Reduce size needed for reusable stubs
Summary: D1776678 increased maxStubSize() from 85 to 134 by adding in space for
an extra instruction I ended up not even needing. I took out the unnecessary
kRipLeaLen and took a couple more opportunities to shrink it:

- kNumServiceReqArgRegs is now 4 instead of 5 since no service requests
  need 5 args.
- We don't emit aligned service requests anymore (since I killed REQ_EXIT) so
  the 15 bytes of padding for alignment are gone.

Each reusable stub will now use 60 bytes. I'm sure we can do even better than
this but this is a good stopping point for now.

I also fixed a bug in MCGenerator::bindJmpccFirst() that was passing a
ConditionCode instead of TransFlags for the last argument to a REQ_BIND_JCC.

Reviewed By: @markw65

Differential Revision: D1817026