Fix relocation assert
Summary: If the target block fills up while relocating, it throws,
and the relocator tries to rewind the work its already done.

There were asserts in the rewind code that were too strong, so
debug builds tended to crash in the relocator when main ran out of
space.

This just makes sure the error is correctly reported as a datablock
filling up.

Reviewed By: @ottoni

Differential Revision: D1818598